Fix API connection by moving types in a different file
