+++ packages/api/src/derives/staking/index.ts
@@ -1,4 +1,4 @@
-export {queryStakingAccountInfo as accountInfo, DerivedStakingInfo} from './stakingAccount';
+export {queryStakingAccountInfo as accountInfo} from './stakingAccount';
 export * from '@polkadot/api-derive/staking/account';
 export * from '@polkadot/api-derive/staking/controllers';
 export * from '@polkadot/api-derive/staking/electedInfo';
